requirements.txt
absl-py==2.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.22
etils==1.3.0
fonttools==4.53.1
glfw==2.7.0
h5py==3.11.0
idna==3.8
importlib_resources==6.4.5
kiwisolver==1.4.7
labmaze==1.0.6
lxml==5.3.0
matplotlib==3.7.5
mujoco==3.2.2
numpy==1.24.4
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
protobuf==5.28.1
PyOpenGL==3.1.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.32.3
scipy==1.10.1
six==1.16.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.1
